refactor(test): rename interactortest to respondertest
